[data][cornell] Observe sequence length for bucket
- #21, need to have bucket, the ideal way would be in each steap, chose one bucket
based on distribution, then randomly pick batch_size elements from this bucket
  - https://mxnet.incubator.apache.org/how_to/bucketing.html
