+++ snowbot/corpus/cornell/dataset.py
@@ -1,11 +1,10 @@
 import os
 import re
 import shutil
-import json
 import pandas as pd
 
 from snowbot.corpus.util import maybe_download, maybe_extract, files_exist, files_missing, train_test_split, \
-    base_vocab_dict
+    gen_vocab
 
 _SEP = '+++$+++'
 
@@ -237,37 +236,3 @@ def is_stupid(line):
         if ('yeah' in line) or ('yes' in line) or ('no' in line):
             return True
     return False
-
-
-def batch_tokenizer(lines, need_clean=False):
-    cleaner = re.compile('(<u>|</u>|\[|\])')
-    vocab_count = {}
-    sentences_tokens = []
-    for line in lines:
-        # TODO: dealing w/ punctuation etc, and do we remove stop words, change he's -> he is etc.
-        if need_clean:
-            line = cleaner.sub('', line)
-        tokens = []
-        for token in line.split():
-            token = token.lower().strip()
-            if token:
-                vocab_count[token] = vocab_count.get(token, 0) + 1
-                tokens.append(token)
-        sentences_tokens.append(tokens)
-    return sentences_tokens, vocab_count
-
-
-def gen_vocab(src, dst, max_words=50000):
-    with open(src, 'r') as f:
-        sentences_tokens, vocab_count = batch_tokenizer(f)
-    print('total words in', src, len(vocab_count))
-    # learned from https://github.com/chiphuyen/stanford-tensorflow-tutorials/blob/master/assignments/chatbot/data.py#L127
-    vocab_sorted = sorted(vocab_count, key=vocab_count.get, reverse=True)
-    vocab = base_vocab_dict()  # pad, unk, s, /s
-    offset = len(vocab)
-    total = min(max_words, len(vocab_count))
-    for i in range(total):
-        vocab[vocab_sorted[i]] = offset + i
-    with open(dst, 'w') as f:
-        json.dump(vocab, f)
-    print('wrote', total, 'words to', dst)
